Renamed libvpr to libarchfpga Added library capability to VPR so that it can be called by another C project Updated Odin II, VPR, and lib reader build environment to new changes
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@446 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ libarchfpga/read_xml_arch_file.c
@@ -485,7 +485,7 @@ static void ProcessPinToPinAnnotations(ezxml_t Parent,
 		annotation->output_pins = my_strdup(Prop);
 		ezxml_set_attr(Parent, "out_port", NULL);
 	} else if (0 == strcmp(Parent->name, "delay_matrix")) {
-		annotation->type = (int) E_ANNOT_PIN_TO_PIN_DELAY;
+		annotation->type = E_ANNOT_PIN_TO_PIN_DELAY;
 		annotation->format = E_ANNOT_PIN_TO_PIN_MATRIX;
 		Prop = FindProperty(Parent, "type", TRUE);
 		annotation->value[i] = my_strdup(Parent->txt);
